Clean up the watchOS example.
